Updated Codebase for slides FE integration

+++ src/main/resources/Database/case_record_schema.sql
@@ -9,8 +9,8 @@ CREATE SEQUENCE IF NOT EXISTS case_record_sequence
 -- Create the case_record table with required columns
 CREATE TABLE IF NOT EXISTS case_record (
     id INT DEFAULT nextval('case_record_sequence') PRIMARY KEY,
-    doctor_id BIGINT NOT NULL,
-    patient_id BIGINT NOT NULL,
+    doctor_id INT NOT NULL,
+    patient_id INT NOT NULL,
     todo VARCHAR(1000),
     date VARCHAR(50) NOT NULL,
     time VARCHAR(50) NOT NULL,
